Improvements to version upgrades
- Add a dedicated Version.cpp source file, improving the separation between version upgrade and core document behavior.
- Minor consistency improvements to the in-progress upgrade logic for MaterialX 1.39.
